# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpoGuide AI — Riyadh Expo 2030 AI Tourist Guide API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Pavilion {
  id: number;
  name: string;
  nameAr: string;
  country: string;
  description: string;
  descriptionAr: string;
  categories: string[];
  zone: string;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  highlights: string[];
  openingHours: string;
}

export interface PavilionWithScore {
  id: number;
  name: string;
  nameAr: string;
  country: string;
  description: string;
  descriptionAr: string;
  categories: string[];
  zone: string;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  highlights: string[];
  openingHours: string;
  relevanceScore: number;
  matchedInterests: string[];
}

export type PavilionStatsCategoryCounts = {[key: string]: number};

export interface PavilionStats {
  totalPavilions: number;
  categoryCounts: PavilionStatsCategoryCounts;
  avgWaitMinutes: number;
  lowCrowdCount: number;
  highCrowdCount: number;
}

export interface Restaurant {
  id: number;
  name: string;
  nameAr: string;
  cuisine: string;
  cuisineAr: string;
  description: string;
  zone: string;
  location: string;
  /** $, $$, or $$$ */
  priceRange: string;
  rating: number;
  /** low, medium, or high */
  crowdLevel: string;
  waitMinutes: number;
  /** @nullable */
  imageUrl?: string | null;
  tags: string[];
  openingHours: string;
}

export interface QueueStatus {
  pavilionId: number;
  pavilionName: string;
  waitMinutes: number;
  /** low, medium, or high */
  crowdLevel: string;
  /** 0-100 crowd capacity percentage */
  crowdPercent: number;
  lastUpdated: string;
  /** increasing, stable, or decreasing */
  trend: string;
}

export interface InterestCategory {
  id: string;
  label: string;
  labelAr: string;
  icon: string;
  description: string;
}

export interface InterestInput {
  /** List of interest category IDs */
  interests: string[];
}

export interface ApiError {
  error: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export type ListPavilionsParams = {
/**
 * Filter by interest category
 */
category?: string;
};

export type ListRestaurantsParams = {
/**
 * Filter by cuisine type
 */
cuisine?: string;
/**
 * Filter by expo zone/area
 */
zone?: string;
};