# lib/api-zod/src/generated/types/geminiImageOutput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpoGuide AI — Riyadh Expo 2030 AI Tourist Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}