# lib/api-zod/src/generated/types/interestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpoGuide AI — Riyadh Expo 2030 AI Tourist Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface InterestInput {
  /** List of interest category IDs */
  interests: string[];
}