# lib/api-zod/src/generated/types/pavilion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpoGuide AI — Riyadh Expo 2030 AI Tourist Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface Pavilion {
  id: number;
  name: string;
  nameAr: string;
  country: string;
  description: string;
  descriptionAr: string;
  categories: string[];
  zone: string;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  highlights: string[];
  openingHours: string;
}