# lib/api-zod/src/generated/types/pavilionStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpoGuide AI — Riyadh Expo 2030 AI Tourist Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { PavilionStatsCategoryCounts } from './pavilionStatsCategoryCounts';

export interface PavilionStats {
  totalPavilions: number;
  categoryCounts: PavilionStatsCategoryCounts;
  avgWaitMinutes: number;
  lowCrowdCount: number;
  highCrowdCount: number;
}