# lib/api-zod/src/generated/types/queueStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpoGuide AI — Riyadh Expo 2030 AI Tourist Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface QueueStatus {
  pavilionId: number;
  pavilionName: string;
  waitMinutes: number;
  /** low, medium, or high */
  crowdLevel: string;
  /** 0-100 crowd capacity percentage */
  crowdPercent: number;
  lastUpdated: Date;
  /** increasing, stable, or decreasing */
  trend: string;
}